BackEnd/MedBotAssist.BotOpenIA/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.2
python-dotenv==1.0.1
langchain==0.3.9
langchain-openai==0.2.9
langchain-community==0.3.9
openai==1.55.3
tiktoken==0.8.0
pydantic-settings==2.6.1
pyodbc==5.2.0
sqlalchemy==2.0.36
python-multipart==0.0.17
PyJWT==2.8.0
httpx==0.28.1
